Extract py dialect ops into veripy/ops_py.py
